Add option to graph only the current state of an entity
This feature adds an option to graph the entity's current state as a
fixed value in the graph rather than graphing the entity's state
history. This feature is useful for displaying fixed values, like
averages, that another entity's state history is compared against.

For example, one might want to plot the current temperature against the
average daily temperature. In this use case, the average daily
temperature is updated once a day, and its state history is not
interesting to compare against. Rather, only its current value is
interesting. The `fixed_value` option allows for the current average
daily temperature to be graphed as a horizontal line.