Support user specifying update and render start date/hour
I will be adding functionality to support the user choosing the range of data to view. This differs from the current latest release, which only allows the user to specify a start date that is a configurable number of hours from the current date and time. I could use some help with the new footer function that displays the timestamps of the current view.