Update usage examples
As more people configure their dashboards entirely through the UI it's becoming an increasingly common error to include a dash at the start of some card configurations, turning it into a list of cards with one entry and causing a confusing syntax error.

It's generally better to avoid `- type:` in usage examples. Especially since they all only contain configuration for a single card.